Fix integer overflows with large files.
HttpsClient.multipartUpload() relied on FileInputStream.available() to
calculate the amount of data to read. Unfortunately, available() can
return a negative value when files are larger than 2GB. The fix is to
simply not use it.

StingleDataSource.open() had an integer overflow in the calculation for
chunkOffset. Added an explicit cast to long.